refactor(dashboards): extract DashboardHeaderActions and DashboardScenePanel
- Extract EditModeActions, FullscreenModeActions, ViewModeActions into
  DashboardHeaderActions component
- Extract ScenePanel (sidebar menu) into DashboardScenePanel component
- DashboardHeader is now a slim orchestrator that composes the extracted
  components
